# docs/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: msmariaulfa33@gmail.com
  version: 0.0.1
  title: DNR Service Swagger Docs
  description: Service for DNR Website
servers:
  - description: Local development
    url: 'http://localhost:3001'
...
paths:
...
  /api/v1/products:
    get:
      summary: Get Products with Filter
      operationId: getProducts
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  totalProduct:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Get array of product by multiple search queries
      parameters:
        - schema:
            type: string
          in: query
          name: category
          description: Get product by category name
        - schema:
            type: string
          in: query
          allowReserved: true
          name: name
          description: Get product by product name
        - schema:
            type: string
          in: query
          name: company_name
          description: Get product by product's company_name
        - schema:
            type: string
          in: query
          name: slug
          description: Get product by product's slug
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
            example: PRICE
            enum:
              - PRICE
              - NAME
              - RATE
              - DISCOUNT
              - TOP_SALES
          in: query
          name: sort_by
          description: Indicating field for sorting the result
        - schema:
            type: string
            enum:
              - ASC
              - DESC
            example: ASC
          in: query
          name: order
          description: Indicating sorting
        - schema:
            type: string
          in: query
          name: min_price
          description: Filtering the minimum price
        - schema:
            type: string
          in: query
          name: max_price
          description: Filtering the maximum price
        - schema:
            type: string
          in: query
          name: rating
          description: Filtering the maximum price
      tags:
        - product
    parameters: []
...
  /api/v1/products/{productId}:
    get:
      summary: Get Product By Id
      operationId: getProductById
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Get single product by Id
      tags:
        - product
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of the product
...
  /api/v1/admin/products:
    get:
      summary: Get Products with Filter for Admin
      operationId: getProductsForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-summary'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Get array of product by multiple search queries
      parameters:
        - schema:
            type: string
          in: query
          name: category
          description: Get product by category name
        - schema:
            type: string
          in: query
          name: name
          allowReserved: true
          description: Get product by product name
        - schema:
            type: string
          in: query
          name: company_name
          description: Get product by product's company_name
        - schema:
            type: string
          in: query
          name: slug
          description: Get product by product's slug
        - schema:
            type: string
            enum:
              - ACTIVE
              - INACTIVE
          in: query
          name: status
          description: Get product by product's status
        - schema:
            type: string
          in: query
          name: page
          description: Get product page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return product
        - schema:
            type: string
            example: PRICE
            enum:
              - PRICE
              - NAME
              - RATE
              - DISCOUNT
          in: query
          name: sort_by
          description: Indicating field for sorting the result
        - schema:
            type: string
            enum:
              - ASC
              - DESC
            example: ASC
          in: query
          name: order
          description: Indicating sorting method
      tags:
        - admin
    post:
      summary: Add a New Product for Admin
      operationId: addProduct
      responses:
        201:
          description: User Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples: {}
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Product-create'
            examples: {}
          text/plain:
            schema:
              type: object
              properties: {}
        description: Post the necessary fields for the API to create a new product.
      description: Add a new product
      tags:
        - product
    parameters: []
  /api/v1/admin/products/upload:
    post:
      summary: Create a New Product for Admin
      operationId: uploadProduct
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Create a new category
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      tags:
        - product
  /api/v1/admin/products/{productId}:
    get:
      summary: Get Product By Id for Admin
      operationId: getProductByIdForAdmin
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      parameters: []
      description: Get single product by Id
      tags:
        - product
    patch:
      summary: Update Product For Admin
      operationId: updateProductById
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Update product by Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product-update'
        description: ''
      tags:
        - admin
    parameters:
      - schema:
          type: string
        name: productId
        in: path
        required: true
        description: Id of the product
  /api/v1/admin/products/{product_id}/upload:
    patch:
      tags:
        - upload
      parameters:
        - in: path
          name: product_id
          schema:
            type: string
          required: true
      summary: upload product photo
      operationId: uploadProductPhoto
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Product-summary'
              examples: {}
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        404:
          description: example error 404 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error404Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
      description: Upload images for product
...
  /api/v1/product_reviews/{productId}:
    get:
      description: Get a product review
      tags:
        - product
      summary: Get a product review
      operationId: getProductReview
      parameters:
        - name: productId
          schema:
            type: string
          in: path
          required: true
        - schema:
            type: string
          in: query
          name: page
          description: Get review page use for pagination
        - schema:
            type: string
          in: query
          name: limit
          description: Limit the return of the review
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  totalPage:
                    type: integer
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product-review'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
  /api/v1/product_reviews:
    post:
      description: Create a new product review
      tags:
        - product
      summary: Create a new product review
      operationId: createProductReview
      security:
        - ApiKeyAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product-review-create'
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product-review'
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
  /api/v1/admin/product_reviews/{reviewId}:
    delete:
      description: Delete a Product Review
      tags:
        - banner
      summary: Delete a Product Review
      operationId: deleteProductReview
      parameters:
        - name: reviewId
          schema:
            type: string
          in: path
          required: true
      responses:
        204:
          description: OK
        400:
          description: example error 400 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error400Response'
        500:
          description: example error 500 response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error500Response'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-auth-token
  schemas:
...
    Error400Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INVALID_API_KEY
            - API_VALIDATION_ERROR
            - CAN_NOT_REGISTER
            - CAN_NOT_LOGIN
            - UNAUTHORIZED_USER
            - OUTLET_CREATION_ERROR
            - CREATE_ORDER_ERROR
            - ADDRESS_NOT_FOUND
            - ADD_TO_CART_ERROR
            - UPDATE_ORDER_ERROR
            - COMPLETE_PAYMENT_ORDER_ERROR
            - CANCEL_ORDER_ERROR
            - COMPLETE_ORDER_ERROR
            - FLASH_SALE_ERROR
            - EXISTING_CUSTOMER_ID
            - RESET_PASSWORD_ERROR
            - NEWS_ERROR
            - OUTLET_TYPES_ERROR
            - PROMOTION_PRODUCTS_EXIST
            - PROMOTION_CODE_EXIST
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
    Error404Response:
      type: object
      properties:
        errorCode:
          type: string
          enum:
            - NOT_FOUND_ERROR
            - USER_NOT_FOUND_ERROR
            - PRODUCT_NOT_FOUND_ERROR
            - ORDER_NOT_FOUND_ERROR
            - BANK_ACCOUNT_NOT_FOUND_ERROR
            - CART_NOT_FOUND_ERROR
            - VERIFICATION_EMAIL_ERROR
            - NOTIFICATION_NOT_FOUND_ERROR
            - FLASH_SALE_NOT_FOUND_ERROR
            - NEWS_ERROR
            - OUTLET_TYPES_ERROR
            - PROMOTION_NOT_FOUND_ERROR
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
      required:
        - errorCode
        - message
    Error500Response:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - INTERNAL_SERVER_ERROR
        message:
          type: string
        errors:
          $ref: '#/components/schemas/ArrayOfErrors'
...
    ArrayOfErrors:
      nullable: true
      type: array
      items:
        oneOf:
          - type: string
          - type: object
...
    Category-summary:
      title: Category
      type: object
      description: Model representing category info for product response
      x-examples:
        Kesehatan:
          value:
            id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
            name: Kesehatan
            icon_url: category/medicine.jpeg
            created: '2019-08-24T14:15:22Z'
            updated: '2019-08-24T14:15:22Z'
      properties:
        name:
          type: string
          description: Name of the category. This value should be unique
        icon_url:
          type: string
          nullable: true
          description: Path to the icon of the category
      required:
        - name
...
    Product-create:
      title: ProductCreate
      type: object
      description: User input schema for adding new product
      x-examples:
        CDR:
          value:
            name: CDR (Calcium-D-Redoxon) ahlinya kalsium suplemen makanan
            sku_number: 000.987.123.456
            company_name: CDR
            description: Suplemen makanan
            unit: PC
            price: 50000
            category_id:
              - Kesehatan
            valid_to: 2019-08-24T14:15:22Z'
      properties:
        name:
          type: string
          description: Product title
        picture:
          type: array
          description: Path to picture file
        sku_number:
          type: string
          description: Product's SKU number
        company_name:
          type: string
          description: Product's brand
        description:
          type: string
          description: Description about the product
        unit:
          type: string
          description: Unit of the product
        price:
          type: integer
          description: Price per unit of product
        discount_percentage:
          type: integer
          description: Percentage of the discount. Fill one of this or the discount_price.
        discount_price:
          type: integer
          description: Discount amount. Fill one of this or the discount_percentage.
        categories:
          type: array
          description: Array of categories name
          items:
            type: string
        valid_to:
          type: string
          description: Last valid date of a product
      required:
        - name
        - sku_number
        - company_name
        - unit
        - price
    Product-summary:
      title: ProductSummary
      type: object
      x-examples:
        CDR:
          value:
            name: CDR (Calcium-D-Redoxon) ahlinya kalsium suplemen makanan
            picture_url: cdr-pic-01
            company_name: CDR
            unit: PC
            slug: cdr-ahlinya-kalsium-suplemen
            price: 50000
      description: Model representing summary information of product for carts & order
      properties:
        id:
          type: string
          description: Id of the product
        name:
          type: string
          description: Product title
        picture_url:
          type: string
          nullable: true
          description: Path to picture file
        slug:
          type: string
        unit:
          type: string
          description: Unit of the product
        price:
          type: integer
          nullable: true
          description: Price per unit of product
        discount_percentage:
          type: integer
          nullable: true
          description: Percentage of the discount
        discount_price:
          type: integer
          nullable: true
          description: Discount amount
        discount_type:
          type: string
          nullable: true
          enum:
            - PRICE
            - PERCENTAGE
            - null
          description: Type of the discount to be shown
      required:
        - name
        - slug
        - unit
...
    Product-update:
      title: ProductUpdate
      type: object
      description: User input schema for updating product
      x-examples:
        CDR:
          value:
            description: Suplemen makanan
            unit: PC
            price: 50000
            category_id:
              - Kesehatan
      properties:
        description:
          type: string
          nullable: true
          description: Description about the product
        unit:
          type: string
          nullable: true
          description: Unit of the product
        discount_percentage:
          type: integer
          nullable: true
          description: Percentage of the discount. Fill one of this or the discount_price.
        discount_price:
          type: integer
          nullable: true
          description: Discount amount. Fill one of this or the discount_percentage.
        price:
          type: integer
          nullable: true
          description: Field to change the price received from SAP.
        categories:
          type: array
          nullable: true
          description: Array of categories name
          items:
            type: string
        status:
          type: string
          nullable: true
          enum:
            - ACTIVE
            - INACTIVE
            - null
        valid_to:
          type: string
          nullable: true
        dpf:
          type: string
          nullable: true
        discount_end_date:
          type: string
          format: date
          nullable: true
...
    Product:
      title: Product
      type: object
      x-examples:
        CDR:
          value:
            id: 497f6eca-6276-4993-bfeb-53cbbbba6f08
            name: CDR (Calcium-D-Redoxon) ahlinya kalsium suplemen makanan
            picture_url: cdr-pic-01
            sku_number: 000.987.123.456
            company_name: CDR
            description: Suplemen manakan
            unit: PC
            slug: cdr-ahlinya-kalsium-suplemen
            price: 50000
            discount_percentage: null
            discount_price: null
            discount_type: null
            status: ACTIVE
            categories:
              - Kesehatan
            valid_to: '2019-08-24T14:15:22Z'
            created: '2019-08-24T14:15:22Z'
            updated: '2019-08-24T14:15:22Z'
      description: Model representing product
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Product title
        picture_url:
          type: string
          nullable: true
          description: Path to picture file
        sku_number:
          type: string
          description: Product's SKU number
        company_name:
          type: string
          description: Product's brand
        description:
          type: string
          nullable: true
          description: Description about the product
        unit:
          type: string
          description: Unit of the product
        slug:
          type: string
          description: Slug url of the product
        price:
          type: integer
          nullable: true
          description: Price per unit of product
        sap_price:
          type: integer
          nullable: true
          description: Price per unit of product
        discount_percentage:
          type: integer
          nullable: true
          description: Percentage of the discount
        discount_price:
          type: integer
          nullable: true
          description: Discount amount
        discount_type:
          type: string
          nullable: true
          enum:
            - PRICE
            - PERCENTAGE
            - null
          description: Type of the discount to be shown
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
          description: Status of the product
        categories:
          type: array
          description: Array of categories name
          items:
            $ref: '#/components/schemas/Category-summary'
        valid_to:
          type: string
          nullable: true
          format: date-time
          description: last valid date for a product
        sold:
          nullable: true
          type: integer
        is_favorite:
          nullable: true
          type: boolean
        is_flash_sale:
          nullable: true
          type: boolean
        dpf:
          nullable: true
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
      required:
        - id
        - name
        - sku_number
        - company_name
        - unit
        - slug
        - status
...
    Product-review:
      title: ProductReview
      type: object
      description: Model representing review for each product
      properties:
        id:
          type: string
          description: Id of the review
        notes:
          type: string
          description: Notes of the review
        rating:
          type: integer
          nullable: true
          description: Rating of the review
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Product-review-create:
      title: ProductReviewCreate
      type: object
      properties:
        notes:
          type: string
          nullable: true
          description: Notes of the review
        rating:
          type: integer
          nullable: true
          description: Rating of the review
        product_id:
          type: string
          description: Product that user reviewed
        order_id:
          type: string
          description: Order that user order the product
...